Add files via upload
## Problem Statement:
🎉🎉 Attention all student data whizzes! We are beyond excited to kick off the "Data Science Student Championship 2023". This year's challenge? A real head-turner: Predicting the 'total_fare' for taxi rides! 🚖

 

As a student, you've undoubtedly hailed a ride when you're late for class or heading out for a late-night study session. Ever wondered how the total cost of your ride is determined? Well, it's time to switch gears from passenger to predictor. In this thrilling hackathon, you'll put your data science skills to the ultimate test by predicting the total fare of a ride-hailing trip. 🎯

 

This championship is designed not just to put your skills under the microscope, but to refine them, strengthen them, and prepare you for real-world problem-solving scenarios. And that's not all! You'll be delving into a fundamental issue in the world of transportation economics.

 
## Meta Data:
* ⏱️ 'trip_duration': How long did the journey last?[in Seconds]
* 🛣️ 'distance_traveled': How far did the taxi travel?[in Km]
* 🧑‍🤝‍🧑 'num_of_passengers': How many passengers were in the taxi?
* 💵 'fare': What's the base fare for the journey?[In INR]
* 💲 'tip': How much did the driver receive in tips?[In INR]
* 🎀 'miscellaneous_fees': Were there any additional charges during the trip?e.g. tolls, convenience fees, GST etc.[In INR]
* 💰 'total_fare': The grand total for the ride (this is your prediction target!).[In INR]
* ⚡ 'surge_applied': Was there a surge pricing applied? Yes or no?