Add files via upload
### Problem Statement: 
Typically a house in Manhattan comprises of various electrical appliances. Each of these appliances partly contributes to the total energy consumption of the house. If these houses are grouped based on their power usage, it helps to analyze the requirement of electric power in different parts of the city. Also, the insights can be useful to identify the appliances with high and low usage of electricity. In this case study, we consider the power consumption due to various appliances in the house and try to group the houses as per their power consumption.

### Data Definition:
The dataset represents the electric power consumption in the house. The data definition is as follows:

Global_active_power: The global minute-averaged active power of the house (in KW)

Global_reactive_power: The global minute-averaged reactive power of the house (in KW)

Voltage: The minute-averaged voltage of the house (in volt)

Global_intensity: The global minute-averaged current intensity of the house (in ampere)

Kitchen_consumption: The power consumption of the appliances in the kitchen (in watt-hour)

Laundry_consumption: The power consumption of the appliances in the laundry room (in watt-hour)

Other_appliances_consumption: The power consumption of the appliances other than in the kitchen and laundry room (in watt-hour)